Prepping AppMan for Mono...
