+++ External/Tools/AppMan/AppMan.csproj
@@ -45,7 +45,7 @@
     <DebugType>none</DebugType>
     <Optimize>true</Optimize>
     <OutputPath>..\..\..\FlashDevelop\Bin\Debug\Tools\appman\</OutputPath>
-    <DefineConstants>FLASHDEVELOP;WIN32</DefineConstants>
+    <DefineConstants>FLASHDEVELOP</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
@@ -74,6 +74,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utilities\ArgProcessor.cs" />
     <Compile Include="Utilities\PathHelper.cs" />
+    <Compile Include="Win32.cs" />
     <Compile Include="ZipLib\BZip2\BZip2.cs" />
     <Compile Include="ZipLib\BZip2\BZip2Constants.cs" />
     <Compile Include="ZipLib\BZip2\BZip2Exception.cs" />
