+++ External/Tools/AppMan/Utilities/ZipHelper.cs
@@ -3,8 +3,6 @@ using System.IO;
 using System.Text;
 using System.Collections.Generic;
 using ICSharpCode.SharpZipLib.Zip;
-using System.Runtime.InteropServices;
-using Microsoft.Win32.SafeHandles;
 using System.Globalization;
 
 namespace AppMan.Utilities
@@ -35,12 +33,10 @@ namespace AppMan.Utilities
                         {
                             Directory.CreateDirectory(dirPath);
                         }
+                        FileStream extracted;
                         String fullPath = Path.Combine(path, entry.Name);
-                        #if WIN32
-                        FileStream extracted = LongPathFileOpen(fullPath);
-                        #else
-                        FileStream extracted = File.Open(fullPath, FileMode.Create, FileAccess.Write);
-                        #endif
+                        if (Win32.IsRunningOnMono) extracted = File.Open(fullPath, FileMode.Create, FileAccess.Write);
+                        else extracted = Win32.LongPathFileOpen(fullPath);
                         while (true)
                         {
                             size = zis.Read(data, 0, data.Length);
@@ -63,57 +59,6 @@ namespace AppMan.Utilities
             // Restore current directory
             Environment.CurrentDirectory = curDir;
         }
-
-        #region LongPath
-
-        // See: http://bcl.codeplex.com/wikipage?title=Long%20Path&referringTitle=Home
-
-        #if WIN32
-
-        private const string LongPathPrefix = @"\\?\";
-
-        [Flags]
-        internal enum EFileAccess : uint
-        {
-            GenericRead = 0x80000000,
-            GenericWrite = 0x40000000,
-            GenericExecute = 0x20000000,
-            GenericAll = 0x10000000,
-        }
-
-        [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
-        internal static extern uint GetFullPathName(string lpFileName, uint nBufferLength, StringBuilder lpBuffer, IntPtr mustBeNull);
-
-        [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
-        internal static extern SafeFileHandle CreateFile(string lpFileName, EFileAccess dwDesiredAccess, uint dwShareMode, IntPtr lpSecurityAttributes, uint dwCreationDisposition, uint dwFlagsAndAttributes, IntPtr hTemplateFile);
-
-        public static FileStream LongPathFileOpen(string path)
-        {
-            string normalizedPath = NormalizeLongPath(path);
-            SafeFileHandle handle = CreateFile(normalizedPath, EFileAccess.GenericWrite, (uint)FileShare.None, IntPtr.Zero, (uint)FileMode.Create, (uint)FileOptions.None, IntPtr.Zero);
-            if (handle.IsInvalid) throw new Exception("Error while resolving path.");
-            return new FileStream(handle, FileAccess.Write, 1024, false);
-        }
-
-        internal static string NormalizeLongPath(string path)
-        {
-            StringBuilder buffer = new StringBuilder(path.Length + 1);
-            uint length = GetFullPathName(path, (uint)buffer.Capacity, buffer, IntPtr.Zero);
-            if (length > buffer.Capacity)
-            {
-                buffer.Capacity = (int)length;
-                length = GetFullPathName(path, length, buffer, IntPtr.Zero);
-            }
-            if (length == 0 || length > 32000)
-            {
-                throw new Exception("Error while resolving path.");
-            }
-            return LongPathPrefix + buffer.ToString();
-        }
-
-        #endif
-
-        #endregion
 
     }
